Added metadata service as opposed to overloading storage
