+++ backend/src/main/java/com/mcmasterbaja/storage/StorageService.java
@@ -50,19 +50,4 @@ public interface StorageService {
    */
   Stream<Path> loadAll(Path dir);
 
-  /**
-   * Reads the headers of a csv file.
-   * @param targetPath The Path of the file to read.
-   * @return A String[] of the headers.
-   */
-  String[] readHeaders(Path targetPath);
-
-  /**
-   * Gets the minimum and maximum values of a column in a csv file.
-   * @param targetPath The Path of the file to read.
-   * @param column The column to analyze.
-   * @return A double[] containing the minimum and maximum values.
-   */
-  Double[] getMinMax(Path targetPath, String column);
-
 }
